+++ src/core/media/engine.h
@@ -113,7 +113,7 @@ public:
     virtual const core::Signal<core::ubuntu::media::Player::PlaybackStatus>& playback_status_changed_signal() const = 0;
     virtual const core::Signal<video::Dimensions>& video_dimension_changed_signal() const = 0;
     virtual const core::Signal<core::ubuntu::media::Player::Error>& error_signal() const = 0;
-    virtual const core::Signal<int64_t>& on_buffering_changed_signal() const = 0;
+    virtual const core::Signal<uint16_t>& on_buffering_changed_signal() const = 0;
 
     virtual void reset() = 0;
 };
